# Faster memoist with less object allocations