#### Summary:
* Track the memoized methods for faster memoize_all/unmemoize_all so we don't have to use introspection methods such as `methods` for information we can "remember"
* Use String interpolation over `Array#join` (less strings, arrays and array operations)
* Use a conditional instead of `Array#compact` (less arrays and array operations)
* Change `escape_punctuation` to be much faster and less allocations for the more common case: methods not ending with `!` or `?`

See each commit for more information and diagnostic scripts to track allocations/time savings.

For a class with 40 memoized methods where we prime and flush the cache 1,000 times.

#### prime_cache/flush_cache with a list of methods:
##### Before:
`ruby prime_flush_cache.rb  0.33s user 0.04s system 98% cpu 0.372 total`

##### After:
`ruby prime_flush_cache.rb  0.19s user 0.03s system 97% cpu 0.221 total`

#### prime_cache/flush_cache with no args:
##### Before:
`ruby prime_flush_cache.rb  0.62s user 0.03s system 99% cpu 0.654 total`

#### After:
`ruby prime_flush_cache.rb  0.21s user 0.03s system 98% cpu 0.243 total`

Additionally, for the same scenario, the object allocations drop:
**58% less objects** for a list of methods pass to prime/flush cache (781025 => 328000 )
**78% less objects** for no args to prime/flush cache (1536562 => 336000)

```ruby
require './lib/memoist'
require 'set'

$methods = Set.new
class Person
  extend Memoist
  1.upto(20) do |n|
    method = define_method("test_#{n}".to_sym) {}
    memoize(method)
    $methods.add(method)
  end

  1.upto(20) do |n|
    method = define_method("test_#{n}?".to_sym) {}
    memoize(method)
    $methods.add(method)
  end
end

p = Person.new
1_000.times do
  p.prime_cache
  p.flush_cache
end
```