@matthewrudy 2015-11-26
Looks good
thanks

(Travis fails on ruby-head, but that seems to be a travis issue)

:+1: